+++ packages/cms-kit/vite.config.js
@@ -4,6 +4,7 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig(() => {
 	return {
 		test: {
+			outputTruncateLength: Infinity,
 			globals: true,
 			coverage: {
 				reportsDirectory: `./coverage`,