+++ packages/create-pantheon-decoupled-kit/src/generators/next-drupal-search-api-addon.generator.ts
@@ -1,4 +1,9 @@
-import { addWithDiff, runLint } from '../actions';
+import {
+	addWithDiff,
+	runLint,
+	runInstall,
+	convertCSSModules,
+} from '../actions';
 import type { DecoupledKitGenerator, DefaultAnswers } from '../types';
 import chalk from 'chalk';
 import { outDirPrompt, tailwindcssPrompt } from '../utils/sharedPrompts';
@@ -22,11 +27,8 @@ export const nextDrupalSearchApiAddon: DecoupledKitGenerator<
 	data: {
 		search: true,
 	},
-	templates: [
-		'next-drupal-search-api-addon',
-		'tailwindless-drupal-search-api-addon',
-	],
-	actions: [addWithDiff, runLint],
+	templates: ['next-drupal-search-api-addon', 'tailwind-shared'],
+	actions: [addWithDiff, runInstall, convertCSSModules, runLint],
 	nextSteps: [
 		`${chalk.cyan(
 			`Run the ${chalk.white.bold(