+++ packages/drupal-kit/src/lib/defaultFetch.ts
@@ -29,8 +29,7 @@ export const defaultFetch = (
 	// and set appropriate cache-control headers on the active response.
 	if (res && typeof res !== 'boolean') {
 		// Ensure api response contains surrogate key headers.
-		headers.set('Pantheon-SKey', '1');
-
+		headers.set('Fastly-Debug', '1');
 		res.setHeader('Cache-Control', cacheControl);
 	}
 
@@ -43,9 +42,9 @@ export const defaultFetch = (
 	// Parse the response to bubble up surrogate keys if possible.
 	fetchPromise
 		.then((response) => {
-			if (res && response.headers.has('Surrogate-Key-Raw')) {
+			if (res && response.headers.has('Surrogate-Key')) {
 				setSurrogateKeyHeader(
-					response.headers.get('Surrogate-Key-Raw'),
+					response.headers.get('Surrogate-Key'),
 					res as ServerResponse,
 				);
 			}